Define class for retrieving secrets
A new class was defined that can be used to retrieve secret values
to use as settings in a Django or Flask website. These secret
values can be retrieved from one of two sources: either AWS
Secrets Manager, or from files that have been encrypted using
systemd-creds. This ensures that Django and Flask websites can
be run using systemd and that sensitive values can be used to
configure the systemd service without exposing them or printing
them in plaintext on the local filesystem.